feat: 커뮤니티 테이블 컴포넌트 분리
